Leverage geometry graph to fix contains.
Unfortunately this entails a `clone` of the geometry. I'm not sure yet
how easy this will be to avoid, since GeomGraph works on Geometry...

This also adds a bunch of undesirable 'static constraints. I think this
is ultimately because of the Box'd LineIntersector. We should try to get
rid of those constraints.